+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.33] - 19 Feb 2018
+### Added
+- unit test to make sure `Model::all()` returns a collection when only only
+  record is retrieved.
+- console command to flush entire model-cache.
+
 ## [0.2.32] - 19 Feb 2018
 ### Fixed
 - hash collision logic to not run query twice if not needed.